+++ experiment/experiment_appveyor_test.go
@@ -19,7 +19,6 @@ func TestMakeDataset_appveyor(t *testing.T) {
 		desc           *datasetDesc
 		dataSourceName string
 		query          string
-		fields         []string
 		want           ddataset.Dataset
 	}{
 		{desc: &datasetDesc{
@@ -28,8 +27,8 @@ func TestMakeDataset_appveyor(t *testing.T) {
 				DataSourceName: "Server=127.0.0.1;Port=1433;Database=master;UID=sa,PWD=Password12!",
 				Query:          "select * from flow",
 			},
+			Fields: []string{"grp", "district", "height", "flow"},
 		},
-			fields: []string{"grp", "district", "height", "flow"},
 			want: dcsv.New(
 				filepath.Join("fixtures", "flow.csv"),
 				true,
@@ -45,7 +44,7 @@ func TestMakeDataset_appveyor(t *testing.T) {
 		BuildDir:      filepath.Join(tmpDir, "build"),
 	}
 	for i, c := range cases {
-		got, err := makeDataset("train", cfg, c.fields, c.desc)
+		got, err := makeDataset(cfg, c.desc)
 		if err != nil {
 			t.Errorf("(%d) makeDataset: %s", i, err)
 		} else if err := checkDatasetsEqual(got, c.want); err != nil {